Add instrumentation to Stresstest
This adds the stacktraces, tracing, sentry and metrics to the stresstest.
The purpose of this is to also measure and profile the effect that various instrumentation has on the performance.
For this, all the various pieces of instrumentation are being initialized with configuration that is as close to real-world as possible, except all the logging and instrumentation is going to the void.
